pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "intelligence-brief"
version = "0.1.0"
description = "AI Intelligence Aggregator - Daily brief for Adam"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "Adam Bell"}
]
dependencies = [
    "httpx>=0.27.0",
    "feedparser>=6.0.10",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "anthropic>=0.40.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "apscheduler>=3.10.0",
    "aiosqlite>=0.19.0",
    "python-dateutil>=2.8.0",
    "tenacity>=8.2.0",
    "structlog>=24.0.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "groq>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
intel-brief = "intelligence_brief.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
